+++ lib/src/_src.g.dart
@@ -7,15 +7,15 @@
 // ▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓
 //.title~
 
-export './not_so_related_to_package/stream_helper.dart';
-export './not_so_related_to_package/date_time_helper.dart';
+export 'not_so_related_to_package/stream_utils.dart';
+export 'not_so_related_to_package/date_time_utils.dart';
 export './not_so_related_to_package/string_to_duration_converter.dart';
 export './type_checking/is_subtype.dart';
 export './type_checking/type_equality.dart';
 export './type_checking/is_nullable.dart';
 export './helpful/deep_get.dart';
 export './converters/let.dart';
-export './extensions/try_or_null_on_function_extension.dart';
+export 'extensions/try_call_on_function_extension.dart';
 export './extensions/value_of_on_enum_extension.dart';
 export './future_or/waiter.dart';
 export './future_or/consec.dart';